Websocket: Remaining types into types.go
These really belong somewhere else mostly, but this will do for now
