--- internal/exchange/websocket/stream_types.go
@@ -1,56 +1,0 @@
-package websocket
-
-import (
-	"time"
-
-	"github.com/thrasher-corp/gocryptotrader/currency"
-	"github.com/thrasher-corp/gocryptotrader/exchanges/asset"
-	"github.com/thrasher-corp/gocryptotrader/exchanges/order"
-)
-
-// PingHandler container for ping handler settings
-type PingHandler struct {
-	Websocket         bool
-	UseGorillaHandler bool
-	MessageType       int
-	Message           []byte
-	Delay             time.Duration
-}
-
-// FundingData defines funding data
-type FundingData struct {
-	Timestamp    time.Time
-	CurrencyPair currency.Pair
-	AssetType    asset.Item
-	Exchange     string
-	Amount       float64
-	Rate         float64
-	Period       int64
-	Side         order.Side
-}
-
-// KlineData defines kline feed
-type KlineData struct {
-	Timestamp  time.Time
-	Pair       currency.Pair
-	AssetType  asset.Item
-	Exchange   string
-	StartTime  time.Time
-	CloseTime  time.Time
-	Interval   string
-	OpenPrice  float64
-	ClosePrice float64
-	HighPrice  float64
-	LowPrice   float64
-	Volume     float64
-}
-
-// UnhandledMessageWarning defines a container for unhandled message warnings
-type UnhandledMessageWarning struct {
-	Message string
-}
-
-// Reporter interface groups observability functionality over Websocket request latency.
-type Reporter interface {
-	Latency(name string, message []byte, t time.Duration)
-}
